add powershell cache keeping logic switch

+++ setup_fedora_wsl_distro.ps1
@@ -1,6 +1,7 @@
 param (
   [switch]$reset = $false,
-  [switch]$remove = $false
+  [switch]$remove = $false,
+  [switch]$keepCache = $false
 )
 
 $fedora_major_version = "33"
@@ -67,7 +68,7 @@ if (!(Test-Path "C:\Temp\fedoraWSL\xz\xz.zip")) {
 }
 
 echo "Downloading Fedora 33 system root..."
-if (!(Test-Path "C:\Temp\fedoraWSL\fedora_root.tar.xz")) {
+if (!(Test-Path "C:\Temp\fedoraWSL\fedora_root.tar.xz") && !(Test-Path "C:\Temp\fedoraWSL\fedora_root.tar")) {
   # Get Fedora system root file
   Invoke-WebRequest -Uri "https://github.com/fedora-cloud/docker-brew-fedora/raw/$fedora_major_version/x86_64/fedora-$fedora_version-x86_64.tar.xz" -OutFile "C:\Temp\fedoraWSL\fedora_root.tar.xz"
 
@@ -83,8 +84,10 @@ if (Test-Path "C:\Temp\fedoraWSL\fedora_root.tar") {
 }
 
 # Clean up
-echo "Cleaning up temporary files..."
-Remove-Item -LiteralPath "C:\Temp\fedoraWSL" -Force -Recurse
+if (!$keepCache) {
+  echo "Cleaning up temporary files..."
+  Remove-Item -LiteralPath "C:\Temp\fedoraWSL" -Force -Recurse
+}
 
 echo ""
 echo "Setup complete!"
